requirements.txt
streamlit
requests
Pillow
opencv-python-headless